+++ allennlp/nn/samplers/sampler.py
@@ -8,5 +8,7 @@ class Sampler(Registrable):
     An abstract class representing a multinomial sampler
     """
 
-    def __call__(self, parameter: torch.Tensor) -> torch.Tensor:
+    def __call__(
+        self, logits: torch.Tensor, num_samples: int = 1, with_replacement: bool = True
+    ) -> torch.Tensor:
         raise NotImplementedError